refactor: move triton vm logic into distinct module
