+++ src/main.rs
@@ -8,66 +8,30 @@ use scalarff::alt_bn128::Bn128FieldElement;
 use scalarff::curve_25519::Curve25519FieldElement;
 use scalarff::foi::FoiFieldElement;
 use scalarff::FieldElement;
-use triton_vm::prelude::*;
 
 mod cli;
 mod compiler;
 mod log;
 mod parser;
+mod provers;
 mod r1cs;
 mod tasm;
 
 fn main() -> Result<()> {
     let mut config = cli::parse()?;
     return match config.target.as_str() {
-        "tasm" => {
-            if config.field != "foi" && config.field != "goldilocks" {
-                return log::error!(
-                    &format!("Unsupported field for target tasm: {}", config.field),
-                    "tasm only support execution in the foi (goldilocks) field"
-                );
+        "tasm" => match provers::tritonvm::prove(&config) {
+            Ok((_stark, _claim, _proof)) => {
+                println!("{:?}", _stark);
+                println!("{:?}", _claim);
+                Ok(())
             }
-            // first compile the assembly
-            let asm = compile_tasm(&mut config)?;
-            // then attempt to prove the assembly in TritonVM
-            let instructions = triton_vm::parser::parse(&asm);
-            if let Err(e) = instructions {
-                return log::error!(&format!("Failed to parse tasm: {:?}", e));
+            Err(e) => {
+                println!("Triton VM errored");
+                println!("{e}");
+                std::process::exit(1);
             }
-            let instructions = instructions.unwrap();
-            let l_instructions =
-                triton_vm::parser::to_labelled_instructions(instructions.as_slice());
-            let program = triton_vm::program::Program::new(l_instructions.as_slice());
-            let public_inputs = PublicInput::from(
-                config
-                    .inputs
-                    .clone()
-                    .into_iter()
-                    .map(|v| BFieldElement::from_str(&v).unwrap())
-                    .collect::<Vec<_>>(),
-            );
-            let secret_inputs = NonDeterminism::from(
-                config
-                    .secret_inputs
-                    .clone()
-                    .into_iter()
-                    .map(|v| BFieldElement::from_str(&v).unwrap())
-                    .collect::<Vec<_>>(),
-            );
-
-            match triton_vm::prove_program(&program, public_inputs, secret_inputs) {
-                Ok((_stark, _claim, _proof)) => {
-                    println!("{:?}", _stark);
-                    println!("{:?}", _claim);
-                }
-                Err(e) => {
-                    println!("Triton VM errored");
-                    println!("{e}");
-                    std::process::exit(1);
-                }
-            }
-            Ok(())
-        }
+        },
         "r1cs" => match config.field.as_str() {
             "foi" => {
                 compile_r1cs::<FoiFieldElement>(&mut config)?;
