+++ src/provers/README.md
@@ -1,0 +1,5 @@
+# provers
+
+Functions to generate proofs and arguments in various types of systems.
+
+Each implementation accepts and compiles an ashlang program, or a compiled [`ar1cs`](./src/r1cs) representation and then generates an argument of knowledge in the selected scheme.
